Add files via upload
Please note that Tetris® is the property of Tetris Holding and this program is not affiliated with any official Tetris other than using the trademark name in the file name.

Also I am not sure how to run these programs. I get them to work sometimes with curl, but yeah, there is no warranty of any kind!